refactor main code into module and update README
