+++ .gitignore
@@ -1,0 +1,8 @@
+# macOS
+.DS_Store
+
+# python
+env/
+__pycache__/
+*.py[cod]
+*$py.class
