+++ README.md
@@ -1,0 +1,66 @@
+# tetris-terminal-game
+
+Tetris Clone.  Runs in terminal.  Implemented in Python using ncurses library.
+
+***IMPORTANT:  MUST HAVE PYTHON3 INSTALLED TO RUN!***
+
+## Setup
+
+To run, type:
+
+```
+python3 tetris.py
+```
+
+## Objective
+
+Clear lines by placing tetris pieces strategically.
+If the pieces stack up to the top, you lose.
+Try to get to level 11!
+
+## Controls
+
+"up arrow" - rotate piece
+"down arrow" - make piece fall faster
+"right arrow" - move piece right
+"left arrow" - move piece left
+"p" - pause game
+"q" - quit game
+
+Thanks for playing!
+
+
+## TO-DO LIST
+
+### Game Mechanics
+
+- wall kicking
+- line completion (sort of)
+- "levels" and associated falling speeds (YES!)
+- score (sort of)
+- "next piece" box (YES!)
+- "retry" option, after losing (YES!)
+- ghosting
+
+### Extras/Aesthetics
+
+- ~~home page~~
+- ~~"lose" mechanism~~
+- ~~"quit" mechanism~~
+- "clear line" animations (Maybe later...)
+- ~~adjust layout/make more clean~~
+- add "juciness" (nope)
+- make screen adjustable? (nope)
+- high score system
+
+### Other
+
+- get tested and advice from other people
+- show andrew as a going-away present!
+-
+
+## Special Thanks
+- carlos caballero
+- jordon wing
+
+## p.s. What was my motivation?  A friend told me that animation in the terminal couldn't be done.  I wanted to prove him wrong, and then some.  So I implemented a full functioning tetris in the terminal. :)
