--- README.txt
@@ -1,30 +1,0 @@
-tetris-terminal-game
-====================
-
-Tetris Clone.  Runs in terminal.  Implemented in Python using ncurses library.
-
-***IMPORTANT:  MUST HAVE PYTHON3 INSTALLED TO RUN!***
-
-to run, type:
-python3 tetris.py
-
-Objective:
-Clear lines by placing tetris pieces strategically.
-If the pieces stack up to the top, you lose.
-Try to get to level 11!
-
-Keyboard inputs:
-"up arrow" - rotate piece
-"down arrow" - make piece fall faster
-"right arrow" - move piece right
-"left arrow" - move piece left
-
-"p" - pause game
-"q" - quit game
-"m" - access menu (includes keyboard input legend)
-
-Thanks for playing!
-Eric Pai,
-Main programmer and designer (Did everything)
-
-p.s. What was my motivation?  A friend told me that animation in the terminal couldn't be done.  I wanted to prove him wrong, and then some.  So I implemented a full functioning tetris in the terminal. :)
